Changed the positions of the planets so they have a kind 3D staggered effect since I can't change the angle of the X plane of the rotation yet.  I don't know how.
I've changed the orbital animation times to use relative values compared to Earth (as the standard) and the time it takes a planet to rotate around the sun.

I've worked out the relative distances from the Sun using mercury's distance as the reference value, thus shrinking Mercury's value reduces the others.

Now I have to change or setup a variable for the transformOrigin of each animation.